lerobot_robot_trlc_dk1/motors/DM_Control_Python/requirements.txt
numpy>=2.0.1
pyserial>=3.5